Betty Style checks

+++ 0x00-hello_world/6-size.c
@@ -2,10 +2,10 @@
 
 /**
  * main - outputs datatype sizes as per computer architecture
- * return: zero on success and non-zero on fail
+ * return: 0  on success and non-zero on fail
  */
-int main (void)
+int main(void)
 {
-	printf("Size of a char: %d byte(s) \n Size of an int: %d byte(s)\n Size of a long int: %d byte(s)\n Size of a long long int: %d byte(s)\n Size of a float: %d byte(s)\n", sizeof(char), sizeof(int), sizeof(long int), sizeof(longlong int), sizeof(float))
+	printf("Size of a char: %d byte(s) \n Size of an int: %d byte(s)\n Size of a long int: %d byte(s)\n Size of a long long int: %d byte(s)\n Size of a float: %d byte(s)\n", sizeof(char), sizeof(int), sizeof(long int), sizeof(long long int), sizeof(float))
 	return (0);
 }
